+++ packages/cra-template-safe-app/template/src/App.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback } from 'react'
 import styled from 'styled-components'
 import { Button, Title } from '@gnosis.pm/safe-react-components'
-import { useSafeAppsSDK } from '@gnosis.pm/safe-apps-react-sdk'
+import { useSafeAppsSDK } from '@safe-global/safe-apps-react-sdk'
 
 const Container = styled.div`
   padding: 1rem;
@@ -47,7 +47,7 @@ const SafeApp = (): React.ReactElement => {
         Click to send a test transaction
       </Button>
 
-      <Link href="https://github.com/gnosis/safe-apps-sdk" target="_blank" rel="noreferrer">
+      <Link href="https://github.com/safe-global/safe-apps-sdk" target="_blank" rel="noreferrer">
         Documentation
       </Link>
     </Container>