+++ packages/safe-apps-test-app/src/App.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import { Spinner, Heading, SegmentedControl } from 'evergreen-ui';
-import SafeAppsSDK, { SafeInfo } from '@gnosis.pm/safe-apps-sdk';
+import SafeAppsSDK, { SafeInfo } from '@safe-global/safe-apps-sdk';
 import { AppTabs } from './types';
 import Main from './tabs/Main';
 import RpcCalls from './tabs/RpcCalls';
@@ -60,7 +60,7 @@ const App = (): React.ReactElement => {
   return (
     <Container>
       <Heading size={700} marginTop="default">
-        Gnosis Safe Test App
+        Safe Test App
       </Heading>
       <SegmentedControl value={currentTab} onChange={(val) => setCurrentTab(val as AppTabs)} options={tabs} />