+++ packages/safe-apps-provider/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @gnosis.pm/safe-apps-provider
 
+## 0.15.2
+
+### Patch Changes
+
+- Updated dependencies [6bdd4f9]
+  - @safe-global/safe-apps-sdk@7.9.0
+
+## 0.15.1
+
+### Patch Changes
+
+- b6c00b8: Convert gaslimit hex-encoded string to a decimal number
+
 ## 0.15.0
 
 ### Minor Changes