+++ packages/safe-apps-wagmi/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gnosis.pm/safe-apps-wagmi
 
+## 2.0.0
+
+### Major Changes
+
+- 147b5e0: This version is bumping versions of wagmi in order to work with wagmi0.10.10
+
 ## 1.2.0
 
 ### Minor Changes